Moved global error handling file
